Initialize icon_path when entry disabled
I am using the jupyter-lmod extension (https://github.com/cmd-ntrf/jupyter-lmod) to dynamically load modules inside a Jupyter container.
The module is able to show/hide icons depending on the availability of the module/executable. For example, this prevents people to try launching RStudio, because they see the icon, if it's not loaded. In this mechanism, the launcher entry has to be set to "disabled" at the beginning (as the module is not loaded). The module takes care of dynamically adding the launcher entry.
However, with the current code, the icon path is not initialized when the module is not enabled from the beginning. And as it's impossible to add more handlers on the same path afterwards (unless I'm mistaken but I tried it multiple ways), the icon can never be displayed.
Proposed modification is to always initialize the icon_path when it's present, even if the entry is disabled. It does not cost much in the end even if this icon/path is never used. And if the entry is there, even if disabled, it must be for a good reason, so why not initialize the icon_path?