Revert addition of set -eux
I think `-e` to error early and `-x` to print statements is the default in a github actions runner already. The addition of `-u` to error on using unset environment variables is the addtion I think isn't a default.

If we need a stricter shell script execution, we can use `defaults.run`, see https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#defaultsrun.