Add server proxy manager
It keeps simpervisor proc objects in internal state to be able to teminate
a process by calling terminate method on proc object.

It also checks the status of active processes by running a periodic
callback. If a process is killed outside of jsp, it removes it from
running apps list.